05-vmarketing-org/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.30,<1
aiosqlite>=0.20,<1
pydantic-settings>=2.6,<3
grpcio>=1.68,<2
httpx>=0.28,<1
python-dotenv>=1.0,<2
alembic>=1.14,<2
pytest>=8.3,<9
pytest-asyncio>=0.24,<1
